#6 Fixing help text containing string 'undefined'…
…. Adding test for that situation, and refactoring test folder
structure while at it. Invoking the testable .mjs file using root
directory as cwd, in order to better simulate real-life situation.
